pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Smorodin_SA <ssmorodin@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.14"
django = "^6.0.2"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.1"
djangorestframework-stubs = "^3.16.8"
django-cors-headers = "^4.9.0"
factory-boy = "^3.3.3"
faker = "^40.5.1"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-django = "^4.12.0"
black = "^26.1.0"
flake8 = "^7.3.0"
pytest-mock = "^3.15.1"
coverage = "^7.13.4"
jsonschema = "^4.26.0"
pytest-xdist = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
